Repaired GUI & Sprites by moving them off GL.Begin().

+++ Assets/Utils/FastNoise/FastNoise.NoiseCubic.cs
@@ -62,8 +62,8 @@ namespace GameEngine.Utils
 			int x3 = x1+2;
 			int y3 = y1+2;
 
-			DECIMAL xs = x-(DECIMAL)x1;
-			DECIMAL ys = y-(DECIMAL)y1;
+			DECIMAL xs = x-x1;
+			DECIMAL ys = y-y1;
 
 			return CubicLerp(
 					CubicLerp(ValCoord2D(seed,x0,y0),ValCoord2D(seed,x1,y0),ValCoord2D(seed,x2,y0),ValCoord2D(seed,x3,y0),
@@ -92,9 +92,9 @@ namespace GameEngine.Utils
 			int y3 = y1+2;
 			int z3 = z1+2;
 
-			DECIMAL xs = x-(DECIMAL)x1;
-			DECIMAL ys = y-(DECIMAL)y1;
-			DECIMAL zs = z-(DECIMAL)z1;
+			DECIMAL xs = x-x1;
+			DECIMAL ys = y-y1;
+			DECIMAL zs = z-z1;
 
 			return CubicLerp(
 				CubicLerp(
